requirements.txt
Django==4.2
djangorestframework
drf-spectacular
python-decouple
djangorestframework-simplejwt
django-redis
redis
ruff
requests
django-ratelimit
django-recaptcha

